refactor(transformer): Moved persistense directory into infrastructure
